backend/requirements_relaxed.txt
python-docx
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python-headless
requests
gdown
reportlab
torch
torchvision
python-pptx
lxml
openai
email-validator
psycopg2-binary
stripe
